requirements.txt
python-dotenv==0.21.0
requests==2.27.1
selenium==4.19.0
undetected-chromedriver==3.5.5
google-api-python-client==2.105.0
capsolver_extension_python==1.0.1